Move Markdown filter into quaddicted tag file. Move quaddicted templatetags to project. Add project templatetags to template settings.
