+++ quaddicted/comments/templates/comments/preview.html
@@ -3,7 +3,6 @@
 {% load comments %}
 {% load static %}
 {% load quaddicted %}
-{% load md2 %}
 
 {% block title %}{% trans "Preview your comment" %}{% endblock %}
